Split out common import routine
